Fix compilation with clang and with GCC -std=c2x
clang -std=c2x still doesn't work because clang is stricter and
doesn't allow K&R-style declarations with -std=c2x.

Also fix a few bugs/clang warnings discovered in the process.

+++ include/region.h
@@ -32,7 +32,7 @@ typedef boolean FDECL((*callback_proc), (genericptr_t, genericptr_t));
 
 struct region_arg {
 	int damage;
-	uchar adtyp;	
+	int adtyp;
 };
 
 typedef struct {
